python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datasynth-py"
version = "1.0.0"
description = "Python wrapper for DataSynth synthetic data generation"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    { name = "EY ASU RnD", email = "michael.ivertowski@ch.ey.com" }
]
maintainers = [
    { name = "EY ASU RnD", email = "michael.ivertowski@ch.ey.com" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing :: Mocking",
    "Typing :: Typed",
]
keywords = [
    "synthetic-data",
    "data-generation",
    "testing",
    "machine-learning",
    "financial-data",
    "accounting",
    "journal-entries",
    "fraud-detection",
]

dependencies = []

[project.optional-dependencies]
cli = ["PyYAML>=6.0"]
memory = ["pandas>=2.0"]
streaming = ["websockets>=12.0"]
pandas = ["pandas>=1.5"]
polars = ["polars>=0.20"]
jupyter = ["jupyter", "matplotlib"]
all = [
    "PyYAML>=6.0",
    "pandas>=2.0",
    "websockets>=12.0",
    "polars>=0.20",
    "jupyter",
    "matplotlib",
]
dev = [
    "PyYAML>=6.0",
    "pandas>=2.0",
    "websockets>=12.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/ey-asu-rnd/SyntheticData"
Documentation = "https://ey-asu-rnd.github.io/SyntheticData/"
Repository = "https://github.com/ey-asu-rnd/SyntheticData"
Changelog = "https://github.com/ey-asu-rnd/SyntheticData/blob/main/CHANGELOG.md"
Issues = "https://github.com/ey-asu-rnd/SyntheticData/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["datasynth_py*"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
